Remove conditional tasks & family-based builds
In GCE, an image family represents the latest image name created.
This enables users to not need to know the exact name, unless required.
Previous to this commit this was leveraged to permit skipping steps
in the VM image build-chain, if relevant files were unmodified.  This
was a mistake in hind-sight.

The VM images need to be intimately related to one another (by their
`$IMG_SFX` value) to prevent clashes if multiple builds are in-flight
concurrently.  Without this, (i.e. relying on image-family) is
ripe for leading to hard-to-debug race conditions.  Worse, these
clashes are not likely to be noticed until much later down the line,
after the images are actually put to use.

For example, one person's experimental PR, could build a new base-image,
while another person's PR builds cache-images.  Some/all of the cash
images could then pick up the experimental base-image builds
unintentionally.

Signed-off-by: Chris Evich <cevich@redhat.com>