Fix label update clash
According to google-cloud support, over the past year, less that 1%
of VM Image label updates clashed with other simultanious lable updates.
When this occurs, an error and message is returned by the gcloud
command:

```
ERROR: (gcloud.compute.images.update) HTTPError 412: Labels fingerprint
either invalid or resource labels have changed
```

Perhaps worse, this causes the label update job in automation to fail,
and require attention from the developer.  Avoid this situation by
detecting this specific condition, issuing a warning, then ignoring it.

This is safe because:

* There are no other automation processes or manual provisions currently
  in use which touch VM Image labels in any way.
* The label update operations are atomic on the cloud-side, this is how
  they are able to detect the error occurring.
* Image lifecycle management automation operates at 1-day granularity.
  Being a few seconds, minutes, or even hours off is inconsequential.

Therefore it is a reasonably safe assumption that another timestamp
update operations caused the clash and the error can simply be ignored.

Signed-off-by: Chris Evich <cevich@redhat.com>