Workaround old packer version in cached image
Cirrus-CI does not invalidate cached images due to ADD/COPY
instructions.  This makes it likely that some changes will
not be reflected in the image_builder container.  Specifically,
updating to a new version of packer.  Fix this by storing
the value in `.cirrus.yml` and reading it out as needed.  This
allows for a single source of truth, at the expense of additional
`Containerfile` complexity.

Signed-off-by: Chris Evich <cevich@redhat.com>