Add action to post image ID to PR on success
It's a common burden to lookup the newly produce VM/Container image ID
produced by automation.  Add a github action which detects Cirrus-CI
'success' and publishes a comment containing a link and the image ID
value.

It's not possible to test this inside a PR here (action code must
exist on master branch).  Test was performed manually here:

https://github.com/containers/automation_sandbox/pull/88

Signed-off-by: Chris Evich <cevich@redhat.com>