[CI:TOOLING] Support the existence of persistent VMs
Further filter the list of orphan VMs to exclude any which have
the `persistent` label and it's value matches `true`.  This must
be done with a negative regex match, as using `!=` operator results
in a warning message being printed.

Also, improve the debug-ability of the script with the addition of
`$DEBUG` and an improvement to `$EVERYTHING` to help in testing the
full filter.  In all cases, testing must still happen manually given
the challenges involved with feeding in mock-data.

Signed-off-by: Chris Evich <cevich@redhat.com>