Reduce image size slightly
Especially for container images, but also for VMs, it's desirable to
not leave a lot of useless junk around.  This is because the images
are compressed and moved around, both of which benefit from a reduced
quantity of bits.  Attempt to improve the situation by selectively
removing/cleaning cached packages and metadata when/where possible.

Signed-off-by: Chris Evich <cevich@redhat.com>