Fix bad update of google-startup-scripts service
The default service unit file is modified at build time to ensure the
Cirrus-CI Agent (runs as a startup script) isn't needlessly restricted
by SELinux policy.  The addition of a comment in a recent version of
the google package broke this build-time modification.  Fix this by
correcting an overly broad regex.

Signed-off-by: Chris Evich <cevich@redhat.com>