Move custom cloud-init file-copy to base-images
For a two-stage image (base + cache) workflow, it makes sense to do as
little as possible in the first stage.  However, when it comes to
cloud-init, somethings are more difficult to undo later.  For example,
default user creation.  For Ubuntu, doing this in the second-stage
(cache) was causing the resulting VM's cloud-init to fail in creating
an 'ubuntu' user.  Since this is undesirable behavior anyway, disable
it universally and relocate the custom override file copy to the first
stage (base).

Also (minor) remove redundant package update command that lists names -
not supported.

Signed-off-by: Chris Evich <cevich@redhat.com>