Workaround broken package update
After spending days debugging, it was determined that upgrading
cloud-init results in a failure to add/update OSLogin users
(and ssh keys) upon reboot (after resetting the system).
The exact cause is unknown and further/deeper debugging
isn't worth the effort.  Workaround the problem by
blocking the update for now