Fix cloud-init service-startup dep. loop
In GCP, there are google services that must start on boot but we also
rely on the more generic cloud-init services to setup the VM.
Specifically, this is necessary to attend to rootfs resizing properly
before testing.  However, there is a dependency loop created when both
are installed, that will result in messages like:

...network-online.target: Found ordering cycle on \
    cloud-init.service/start
...network-online.target: Found dependency on \
    google-network-daemon.service/start
...network-online.target: Found dependency on \
    google-instance-setup.service/start
...network-online.target: Found dependency on \
    network-online.target/start
...network-online.target: Job cloud-init.service/start \
    deleted to break ordering cycle starting with \
    network-online.target/start

Previously, this was addressed by depositing a customized
cloud-init.service file for Fedora VMs.  However, at some point in the
past this step was dropped.  Restore it.

Also, add a header-comment to this and another custom service file to
indicate where they came from.

Signed-off-by: Chris Evich <cevich@redhat.com>