Multiple pythons needed for podman-py testing
The one-size-fits-all approach is ever increasingly showing gaps and
inefficiencies WRT CI for the `podman-py` repository.  Rather than
continue to increase the bloat of the golang-centric images, produce a
dedicated image for this repo.  This allows pre-installing/caching all
the necessary python bits needed, avoiding the need to install them at
runtime.  It will require updates to the `podman-py` setup in order to
make use of the new image effectively.

Signed-off-by: Chris Evich <cevich@redhat.com>