Drop Fedora 35 runc/CGroupsV1, enable in Ubuntu
Prior to this commit, Ubuntu VMs were setup with CGroupsV2 enabled.
This supported (especially) podman testing to occur with `crun` instead of
`runc` + CgroupsV1 which is the default.  The opposite was then done for
the Fedora N-1 (aka 'prior') VMs, CGroupsV1 was enabled + `runc`.

With the change in F36 to a netavark/aardvark-dns network stack from
CNI, testing matrices will become additionally complicated if this
regime was maintained.  Simplify the landscape by reverting the Ubuntu VM
setup back to it's default of CGV1 + `runc`.  At the same time, restore
Fedora 35 to it's default of CGV2 + `crun` (along with Fedora 36).

In other words, coverage will apply according to distribution defaults:

* Latest Ubuntu: Use CGV1 + `runc` + CNI
* Prior Fedora: Use CGV2 + `crun` + CNI
* Latest Fedora: Use CGV2 + `crun` + netavark/aardvark-dns

Signed-off-by: Chris Evich <cevich@redhat.com>