[CI:TOOLING] Fix runtime check of non-exported variable
Related to https://github.com/containers/automation_images/pull/123 the
installer checks that it's running under an expected environment.  This
is always the case under test, but not under normal runtime conditions.
Fix this by making sure the check-variable is available to the installer
when auto-updating.

Also, update CI config so test-build-push task uses the latest
build-push image by family-name instead of the image tied to the PR.
This allows testing to occur under both normal and (magic)
`[CI:TOOLING]` PRs.

Signed-off-by: Chris Evich <cevich@redhat.com>