Add support for reading PR labels
On occasion it's necessary to alter automation behavior based on some
criteria.  For high-level conditions, using magic-words in the PR title
is quick/dirty but somewhat cumbersome.  Using the `required_pr_labels`
mechanism in `.cirrus.yml` is handy too, unless the 'absence' of a label
needs to be tested.  Add a generic library function that allows retrieving
the labels on a PR.  Future commits can make use of this in scripts to
take various actions based on presence or absence of the labels.

Signed-off-by: Chris Evich <cevich@redhat.com>