Add support for reading PR labels
On occasion it's necessary to alter automation behavior based on some
criteria.  For high-level conditions, using magic-words in the PR title
is quick/dirty but somewhat cumbersome.  Using the `required_pr_labels`
mechanism in `.cirrus.yml` is handy too, unless the 'absence' of a label
needs to be tested or some combination of multiple labels.

Add a generic library function that allows retrieving the labels on a
PR.  However it requires the automation-tooling library to be loaded.
Add this to the image-builder container and VM image. These environments
will be where `get_pr_labels()` will most likely be used.

Signed-off-by: Chris Evich <cevich@redhat.com>