Build from stock Fedora GCE provided images
Supporting a new base-image in downstream environments is often
a time-consuming and difficult.  For Fedora, normally this work begins
when a release goes into beta.  However, beta versions of Fedora
releases are not provided in google, they must be specially imported.
Otherwise, after release the `fedora-cloud` project provides read-made
images that can be used in a vastly simpler manner.

Alter the base-image build workflow to use the pre-built images by
default.  However if a PR has the 'Beta' label set, cause the workflow
to import the development fedora images instead.

Signed-off-by: Chris Evich <cevich@redhat.com>