Fix podman_tooling script
This script is needed by both podman and buildah CI.  However a bug in
the line `export GOPATH="${GOPATH:/var/tmp/go}"` was causing it to
silently exit before doing it's job.  Fortunately the downstream build
systems are somewhat tolerant, so the problem wasn't noticed at image
build time.

Simplify the script by requiring it to be called under $SUDO, this
allows moving a bunch of golang variable definitions into the script and
also making it `set -e` so failures are noticed sooner.

Signed-off-by: Chris Evich <cevich@redhat.com>