Skip specified image builds based on PR labels
With an ever growing number of images to build, it's inevitable that
a flake will occur on an image the author does not care about.  Should
this occur, provide a runtime-method by which a 'no_<name>' label can be
added to the PR before re-running the failure.  The specified `<name>`
build will then be bypassed.

Note: This will not bypass container tooling images (they're always
required). Nor will it properly handle any build dependencies.
For example a `no_fedora` label will cause a build failure of the
`build-push` image, since it depends on the `fedora` base image.

Signed-off-by: Chris Evich <cevich@redhat.com>