Allow some images test-timestamp updates to fail
Normally (in other CI systems) we want to know immediately if updating
timestamps fails.  Otherwise an image could be pruned due to old-age
without anybody noticing.  However on this specific repository, it's
possible that not all images are actually built.  So it's necessary
to allow for some (possibly all) to fail.  However, this shouldn't
impact the testing of the `imgts` container & scripts in CI (here).

Signed-off-by: Chris Evich <cevich@redhat.com>