Support optional EC2 images in imgts container
AWS EC2 has a separate mechanism for recording an AMI's last-used
timestamp.  However, other important metadata is also maintained on
images such as `permanent=true` (or not).  For repos. which utilize EC2,
add support for the imgts container updating EC2 images metadata.

Also, add a task to exercise the imgts container in a [CI:TOOLING]
context (where the image is built).  Set this up as a dry-run
so it doesn't depend on the images actually existing.

Signed-off-by: Chris Evich <cevich@redhat.com>