Add EC2 imgobsolete container & cirrus-cron support
Every time images are built in this repo. it can be a dice-roll.
There's a decent possibility the build could fail or the images won't
meet the needs of downstream CI.  Additionally, when downstream CI
actually moves onto updates images, the old ones are left behind taking
up space.  Multiply this by all the repos using VMs, and it quickly
adds up to a lot of crap images hanging around.

Dealing with this mess is the job of the imgobsolete container image.
Add support to it for handling AWS EC2 images, including the required
special-handling for (release-branch) permanent=true images. Also update
the `test_imgobsolete` task to validate the changes actually functional.
Lastly, update the two related task-names to make them easier to
read/understand.

Signed-off-by: Chris Evich <cevich@redhat.com>