Install aws CLI version > 1.22 in imgts container
The `imgts` image is a base-image for many other tooling images.
In one of these, a future commit will need to use the aws CLI to
query the `LastLaunchedTime` AMI attribute.  This operation is
not supported in the `awscli` packaged shipped with CentOS8, or
F36 for that matter. Install the latest version from upstream
when building the `imgts` container.

Signed-off-by: Chris Evich <cevich@redhat.com>