Fix silent imgobsolete error + check status JSON
Fixes: #186

For whatever reason, the `aws` CLI does not exit non-zero if there was
an error obsoleting an image.  Worse, it forces callers to verify a JSON
blob as well.

Fix this, and improve the script by accumulating any errors and exiting
non-zero after the entire loop finishes.  In the case of `aws`,
additionally verify the output JSON provides positive confirmation of a
successful update.

Signed-off-by: Chris Evich <cevich@redhat.com>