Workaround packer amazon_import post-processor bug
Ref: https://github.com/hashicorp/packer-plugin-amazon/issues
     number 264

Implement a rough and ready set of make targets as a stand-in for
the broken packer amazon_import post-processor.  This works, but
is fairly fragile:  There are several failure modes which cannot
easily be checked automatically.  For example, the s3 upload can
silently fail and/or store corrupted data.

When the packer bug is resolved, this commit can most probably be
safely  reverted.  This should make the process entirely automated
(via the `automation_images` PR workflow), possibly with only
minimal updates.

Add a `import_images/README.md` reference for the semi-manual
process.

Signed-off-by: Chris Evich <cevich@redhat.com>