Lookup Fedora image & checksum URLs on demand
Previously, maintainers were required to not only fill out the desired
Fedora release numbers, but also manually look up six separate URLs and
copy-paste their values.  This is a pain in the butt and error-prone.
Add a script to perform the lookups and call the script for the required
URLs only when required.  In other words, don't execute the script for
`make help` or other targets that don't require the URLs.

Also update import_images documentation to reflect the URL-update
requirement, and tweak several other minor details.

Signed-off-by: Chris Evich <cevich@redhat.com>