Ensure CNI testing in Fedora N-1 CI VM Images
Unfortunately there are still some corner use-cases for CNI even in
supported Fedora versions.  In the prior iteration of images, release 35
was hard-coded to force CNI use.  That was immediately bypassed when
updating to F36/37 releases.  Fix the condition to check for the word
"prior" in the build name, so CNI is always used on Fedora N-1.

Signed-off-by: Chris Evich <cevich@redhat.com>