Add FLAVOUR build-arg support to build_push
This is needed to support image-builds which rely on passing a build-arg
to indicate the flavour-name or type of image.  For example, `upstream`,
`testing`, or `stable`.  Accommodations are also made to continue
supporting the previous method - where the flavour is taken as the
base-name of the context directory.  To prevent any ambiguity, the
flavour argument will be preferred over the sub-directory name.

In all cases, the flavor value will be passed to buildah via
a `--build-arg FLAVOUR=<value>`.  Additionally it will
be passed to the `tag-version.sh` mod-script as the first argument,
with the command version number as the second argument.

Signed-off-by: Chris Evich <cevich@redhat.com>