Fix ssh disconnect during package update
As originally found in:
Ref: https://github.com/containers/automation_images/issues/243

Due to memory requirements needed to process an ever-growing updates
repo., the `dnf` process is being OOM killed.   Ref:

https://bugzilla.redhat.com/show_bug.cgi?id=1907030

Since this is repo-content bound, it will eventually afflict F37 as well.

Temporarily workaround the issue by disabling the updates repo when
building base-images.  While this is kicking the can down the road,
the environment where cache-images build has more memory available
(and is also easier to debug).  Also pre-build the package cache outside
of any package transactions to further reduce memory requirements (as per
comments in the bug above).

Signed-off-by: Chris Evich <cevich@redhat.com>