Modify windows image to pre-create an interactive session.
This is used to support tunneling commands from headless session 0 to an active
logged-in interactive session, as session #1.

The session is created by Winlogon autologon, which is enabled by the
auto_logon.ps1 script. Additionally, psexec is installed to support launching
through a new session and auth via the Service Control Manager.

Signed-off-by: Jason T. Greene <jason.greene@redhat.com>