Validate: Catch attempted re-use of IMG_SFX
IMG_SFX values need to change for every image build, even within the
same PR.  Attempt to catch re-use of a suffix before starting the
lengthy build process (which will eventually fail on a duplicate).
Check the `imgts` image simply because it builds very early in
cirrus-ci and cannot be skipped with a "no_*" PR label.

Also - enforce IMG_SFX change rules when building in `[CI:TOOLING]`
mode, since those container images push with a IMG_SFX tag.  It would be
unfortunate if an unrelated PR clobbered in-use tooling images w/ a
breaking-change.

Signed-off-by: Chris Evich <cevich@redhat.com>