Fix perpetually re-deprecating AMIs
Previously while searching through AMI's to label and/or mark for
deprecation the imgobsolete container never checked if the AMI was
already deprecated.  Since the deprecation date is based on the current
date/time, this results in deprecated AMI's never actually being
deprecated.  Fix this by skipping AMIs that have a deprecation date
already set.

Signed-off-by: Chris Evich <cevich@redhat.com>