Automatically select docker download Debian repo.
When Debian CI VM Images were initially implemented, there was no Docker
repository for the Debian SID release, so the latest release code-name
was hard-coded.  The situation has changed, and there is now a Docker
repository for bookworm.  Remove the hard-coded release code-name and
replace it with a name-lookup from `/etc/os-release` (currently
`bookworm`).

Signed-off-by: Chris Evich <cevich@redhat.com>