Automatic termination of EC2 VMs
Around the time of this commit, an annoyingly steady stream of EC2
orphans were reported to Cirrus-support.  They've taken actions to
resolve, but the failure-modes are many and complex.  Since most of the
EC2 instances are rather expensive to keep needlessly running, and manual
cleanup is annoying, enhance the monitoring script try to attempt
termination automatically.

This isn't perfect, it's possible for the script to break in strange ways
and it's not practical to check for all of them.  Instead, include
some helpful indications in the monitoring e-mail regarding what was
attempted.

Signed-off-by: Chris Evich <cevich@redhat.com>