Build-Push: Install pip and wait-for-copr
Podman-Desktop team would like fcos images built with packages from
rhcontainerbot/podman-next for MacOS testing with the latest unreleased
bits.
Ref: https://github.com/containers/podman/issues/19448

This commit installs pip and wait-for-copr in the build-push images.
wait-for-copr, as the name suggests, waits for a build with a specified
string to become available on a copr repo.
Ref: https://github.com/packit/wait-for-copr

Signed-off-by: Lokesh Mandvekar <lsm5@fedoraproject.org>