Obscure non-cache image IDs in pr-comment table
All built images are included in the build-table added as a PR comment
to be helpful for reference and possible debugging.  However, it's
unhelpful if a human accidentally tries to deploy a non-cache image ID
into CI somewhere.  Those images are never to be used outside of very
special-case situations.  Obscure non-cache image IDs in the table to
prevent accidents.

Signed-off-by: Chris Evich <cevich@redhat.com>