Update automation-library
Significantly, this version defines a `passthrough_envars()` function to
replace the two duplicate definitions in podman and buildah CI.  When
incorporating the new images into those environments, the duplicates
should be removed.

Also included is an important updates to the build-push script that
improves debugging in cases where the `--nopush` argument is used.

Signed-off-by: Chris Evich <cevich@redhat.com>