Implement quay.io container image build and push
This job use to be performed from the individual repositories CI,
however there was a major problem:

https://github.com/containers/podman/discussions/19796

Reinstate the build jobs in this repo. since it's secrets are secure and
builds are safe from general-public meddling.

Also, slighly alter the existing cirrus-cron triggered tasks such that
they only respond to a specific job name.

Signed-off-by: Chris Evich <cevich@redhat.com>