Update windows CI VMs for hyper-v machine testing
In addition to updating mingw and golang, this moves the
installation of .Net and wixtoolset here instead of at CI runtime.
The windows packer-configuration was updated to operate more
consistently with how things are done in Linux WRT calling scripts.
Along with some file renames and other cosmetic changes, the windows
build timeout was increased since the extra packages seem to
place it right on the edge of the former value.

Signed-off-by: Chris Evich <cevich@redhat.com>