Attempt to fix URL in notification mail
The docs are not specific enough to know for sure `run_id` is the
correct value to use.  When browsing to a job, there are two numbers
present in the URL, I cannot find a ref for the one of them :S
Hopefully `run_id` is correct and the second number isn't needed.

Signed-off-by: Chris Evich <cevich@redhat.com>