Support no-clone build-push mode
As originally conceived, the build context for each image lives in the
respective podman, buildah, and skopeo repositories.  A future set of
PRs will move both the source and build automation into the
new containers/image_build repository.  This is needed to support
images that are point-in-time rebuildable and run test-builds on
image context changes.

Add a magic 2nd argument prefix ('.'), and conditionals to prevent
cloning the build context repo. This will allow for an interim period
where build automation can run from both the current and new repository
until the context repos can be moved.

Signed-off-by: Chris Evich <cevich@redhat.com>