Migrate build script from c/automation_images
Ref: https://github.com/containers/image_build/pull/12

Previously the build-push scripts were run by automation in this repo.
That has since changed, with a migrated over to the
containers/image_build repo.  However, while automation there uses the
most recent build-push VM image, that image is produced in this repo.
Arrange to test the latest script against just-produced VM images to
ensure the environment is always supportive for the script.

Signed-off-by: Chris Evich <cevich@redhat.com>