Simplify latest image tagging
Previously when a PR was merged, another build ran for all the critical
container images, along with tagging them 'latest'.  This is not ideal,
because the content can change from the time the PR build and tested the
images until when it was merged.  There is also an anticipated future
need to access the `fedora_podman` and `prior-fedora_podman` images via a
"latest" tag.

* Update image-build tasks to only run in PRs
* Simplify `ci/make_container_images.sh` to no-longer require/use a
  magic `$PUSH_LATEST` value.
* Deduplicate all FQIN references to reuse a common prefix in `$REGPFX`
* Add a new `tag_latest_images_task` that only runs on branches, and
  simply adds a `latest` tag to all container images based on the
  (as-merged) value of `$IMG_SFX` (from IMG_SFX file)

Signed-off-by: Chris Evich <cevich@redhat.com>