Yet another new Pasta (04-05)
This one fixes a user-reported bug that we don't see in CI.
It's in bodhi for rawhide but no others. We want to test anyway.

Also, small changes to Windows Chocolatey install command
to conform to (some) best practices document. Link to such,
and explain why I disregard some of what they call "best".

Signed-off-by: Ed Santiago <santiago@redhat.com>