Simplify setting Debian release version
Previously a convoluted system was used to add a "fake" release number
into `/etc/os-release` for CI/automation purposes.  It forced a
two-component version to satisfy some legacy automation-library needs.

Since the release number is also specified in the Makefile, and passed
into the packer call, it's trivial to simply provide this value to the
`debian_base-setup.sh` script.

This reduces complexity and avoids duplication.

Signed-off-by: Chris Evich <cevich@redhat.com>