Fix tagging latest after [CI:TOOLING] PR merge
After a PR merges a branch-level job runs to tag the new container
images.  However, there is a special-case when a magic string is present
in the PR title:  No Fedora/Skopeo images were be built, so they should
not be tagged be ignored.

Prior to this commit, this special case isn't handled correctly, because
`CIRRUS_CHANGE_TITLE` only contains the first-line of the HEAD commit.
When executing on a branch, after a PR merge, this would be something
like:

`Merge pull request #FOO from some/thing`

Therefore not matching the intended magic string.  Fix this by switching
to a check against `CIRRUS_CHANGE_MESSAGE` which includes the entire
message.  Importantly, when merged using the github UI, the second line
of the commit message should contain the PR description and thus the
magic string.

Signed-off-by: Chris Evich <cevich@redhat.com>