Fix renovate updating lib.sh
Previously Renovate was failing in a multi-line search for an anchored
pattern in `lib.sh`.  This resulted in it completely ignoring the custom
regex manager for that file, as observed in the debug logs.  Fix this by
removing the regex anchors.

Also remove the filename anchors referenced in the `lib.sh` package rule
as they're unnecessary.

Signed-off-by: Chris Evich <cevich@redhat.com>