add bpftrace for CI debugging
I like to run a bpftrace based program in CI to collect better logs for
specific processes not observed in the normal testing such as the podman
container cleanup command.

Given you need to have full privs to run ebpf and the package pulls in
an entire toolchain which is almost 500MB in install size we do not add
it the the container images to not bloat them without reason.

https://github.com/containers/podman/pull/23487

Signed-off-by: Paul Holzinger <pholzing@redhat.com>