windows: add zstandard package
Windows does not have zstd by default so we need to install it. In
particular I am looking at switching the repo archive to zstd as this
makes things much faster (over 1min in podman)[1] but the windows
testing is unable to extract that. While archiver added zstd support a
while back it is not in the version that is on chocolatey which seems a
bit out of date.

[1] https://github.com/containers/podman/pull/24120

Signed-off-by: Paul Holzinger <pholzing@redhat.com>