skopeo / fedora cache_image: Install docker-distribution
Having the registry binary named `registry-v2` causes trouble for
`make test-integration-local`. The registry binary provided by the
docker-distribution package is just `/usr/bin/registry`.

Depending on docker-distribution should make things simpler, more
consistent and usable regardles of CI / testing environment.

In skopeo cirrus jobs, the integration tests are run on the host itself
but a lot of the binaries are copied from the skopeo_cidev container.
So, in this case docker-distribution is directly installed on the host
environment and the registry-v2 build is removed from the skopeo_cidev
image.

Signed-off-by: Lokesh Mandvekar <lsm5@redhat.com>